Move query_crlite to scan subpackage, remove demo code
